requirements.txt
pdf2image==1.16.3
Pillow==10.0.0
Flask==2.3.3
Werkzeug==2.3.7
rembg==2.0.50
numpy>=1.23.5
PyPDF2==3.0.1
